+++ pkgs/app/src/components/Flow/Details/index.module.scss
@@ -8,7 +8,7 @@
   display: flex;
   flex-direction: column;
 
-  height: calc(100vh - $height-header-project-total);
+  height: 100vh;
 
   background-color: #fff;
