+++ pkgs/app/src/components/NotFound/index.module.scss
@@ -10,3 +10,7 @@
   flex-direction: column;
   align-items: center;
 }
+
+.action {
+  margin-top: var(--spc-2xl);
+}
