+++ pkgs/app/src/components/NotFound/index.tsx
@@ -1,6 +1,7 @@
 import { Helmet } from 'react-helmet-async';
 import { Link } from 'react-router-dom';
 
+import { useProjectStore } from '../../common/store';
 import { titleSuffix } from '../../common/string';
 import { Button } from '../Form/Button';
 
@@ -10,6 +11,7 @@ export const NotFound: React.FC<{ title?: string; message?: string }> = ({
   title = 404,
   message = 'Sorry, the page you visited does not exist.',
 }) => {
+  const { project } = useProjectStore();
   return (
     <div className={cls.notfound}>
       <Helmet title={`404 ${titleSuffix}`} />
@@ -17,7 +19,10 @@ export const NotFound: React.FC<{ title?: string; message?: string }> = ({
         <h2>{title}</h2>
         <p>{message}</p>
 
-        <Link to="/">
+        <Link
+          to={project ? `/${project.orgId}/${project.slug}` : '/'}
+          className={cls.action}
+        >
           <Button display="primary">Back Home</Button>
         </Link>
       </div>
