+++ pkgs/app/src/components/Toast/index.module.scss
@@ -3,7 +3,7 @@
 .portal {
   position: fixed;
   z-index: 9989892;
-  top: $height-header + 16px;
+  top: 16px;
   right: 16px;
 
   display: flex;
