+++ pkgs/app/src/styles/theme.scss
@@ -105,12 +105,5 @@ $main-col-w: 780px;
 $min-w: 1024px;
 $max-w: 1216px;
 
-// Height
-$height-header: 46px;
-$height-header-org: 76px;
-$height-header-org-total: $height-header + $height-header-org;
-$height-header-project: 76px;
-$height-header-project-total: $height-header + $height-header-project;
-
 // Transition
 $trst-smooth: cubic-bezier(0.2, 0, 0, 1);
