+++ pkgs/app/src/views/Homepage/index.module.scss
@@ -4,5 +4,5 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  height: calc(100vh - $height-header);
+  height: 100vh;
 }
