+++ pkgs/app/src/views/Project/Deploy/List/index.module.scss
@@ -2,7 +2,7 @@
 
 
 .header {
-  padding: $p-3xl $p-5xl;
+  padding: var(--spc-6xl);
 }
 
 .list {
@@ -11,7 +11,7 @@
 }
 
 .row {
-  padding: $p-l $p-5xl $p-xl;
+  padding: var(--spc-l) var(--spc-6xl) var(--spc-xl);
   border-bottom: 1px solid var(--border1);
 
   .title {
