+++ pkgs/app/src/views/Project/Revisions/List/index.module.scss
@@ -6,7 +6,7 @@
   align-items: center;
   justify-content: space-between;
 
-  padding: $p-3xl $p-5xl;
+  padding: var(--spc-6xl);
 
   .placeholderReset {
     pointer-events: none;
@@ -20,7 +20,7 @@
 }
 
 .row {
-  padding: $p-l $p-5xl $p-xl;
+  padding: var(--spc-l) var(--spc-6xl) var(--spc-xl) var(--spc-6xl);
   border-bottom: 1px solid var(--border1);
 
   .title {
